Structs

ManifoldAPI

ManifoldAPIs.ManifoldAPI — Type

ManifoldAPI <: AbstractAPI

Basic object for Manifold API

Fields

api_url::String: base url for API

Market

ManifoldAPIs.Market — Type

Market

An abstract market type.

Multiple

ManifoldAPIs.Multiple — Type

Multiple <: Multiple

An abstract type for markets with multiple outcomes

MultipleLinked

ManifoldAPIs.MultipleLinked — Type

MultipleLinked <: Multiple

A set of multiple related market with prices that sum to 1.

MultipleUnlinked

ManifoldAPIs.MultipleUnlinked — Type

MultipleUnlinked <: Multiple

A set of multiple related market with prices that are not required to sum to 1.

Single

ManifoldAPIs.Single — Type

Single <: Market

A market for a single binary outcome.

Get Functions

The functions described in this section allows the user to request market information from Manifold.

get_all_markets

ManifoldAPIs.get_all_markets — Function

get_all_markets(api::AbstractAPI)

Returns information for 1000 prediction markets.

Arguments

api::AbstractAPI: API object

Example

using ManifoldAPIs

api = ManifoldAPI()
markets = get_all_markets(api)

get_market_by_slug

ManifoldAPIs.get_market_by_slug — Function

get_market_by_slug(api::AbstractAPI, market_slug)

Returns prediction market information by slug.

Arguments

api::AbstractAPI: API object

market_slug: market descriptor e.g., supreme-court-rules-trump-tariffs-u in https://manifold.markets/QuimLast/supreme-court-rules-trump-tariffs-u

Example

using ManifoldAPIs

api = ManifoldAPI()
market_slug = "supreme-court-rules-trump-tariffs-u"
market = get_market_by_slug(api, market_slug)

get_market_price

ManifoldAPIs.get_market_price — Function

get_market_price(api::AbstractAPI, market_id)

Returns market price based on market id.

Arguments

api::AbstractAPI: API object

market_id: unique market id

Example

using ManifoldAPIs
api = ManifoldAPI()
get_market_price(api, "9t61v9e7x4")

get_bets

ManifoldAPIs.get_bets — Function

get_bets(api::AbstractAPI;
    market_id = "",
    user_name = "",
    market_slug = "",
    limit  = "",
    before  = "",
    after  = "",
    before_time  = "",
    after_time  = "",
    kinds  = "",
    order  = "",
)

Returns a vector of bets or trades for a given market id and/or user.

Arguments

api::AbstractAPI: API object

Keywords

market_id = "": unique market id

username = "": username associated with a trader

market_slug = "": market descriptor e.g., supreme-court-rules-trump-tariffs-u in https://manifold.markets/QuimLast/supreme-court-rules-trump-tariffs-u

limit = "": the number of bets to return. The default and maximum are both 1000.

before = "": Include only bets created before the bet with this ID. For example, if you ask for the most recent 10 bets, and then perform a second query for 10 more bets with before=[the id of the 10th bet], you will get bets 11 through 20.

after = "": Include only bets created after the bet with this ID. For example, if you request the 10 most recent bets and then perform a second query with after=[the id of the 1st bet], you will receive up to 10 new bets, if available.

before_time = "": include only bets created before this timestamp.

after_time = "": include only bets created after this timestamp.

kinds = "": Specifies subsets of bets to return. Possible kinds: open-limit (open limit orders, including ones on closed and reolved markets).

order = "": asc or desc (default). The sorting order for returned bets.

Example

using ManifoldAPIs
api = ManifoldAPI()
get_bets(api; market_slug = "supreme-court-rules-trump-tariffs-u")

Post Functions

make_bet

ManifoldAPIs.make_bet — Function

make_bet(api::AbstractAPI, header, order)

Completes a transaction for a specified amount and market. Details about the API can be found at https://docs.manifold.markets/api#post-v0bet

Arguments

api::AbstractAPI: API object

header: a dictionary or other compliant data structure containing authorization details

order: a dictionary or other compliant data structure containing details of transaction.

Example

using ManifoldAPIs

header = Dict(
    "Authorization" => "Key authorization_key_here",
    "Content-Type" => "application/json"
)

order = Dict(
    "contractId" => "HtsxyBopv0MSywM0f0Yp",
    "amount" => 10,
    "outcome" => "YES",
    "dryRun" => true
)

api = ManifoldAPI()
make_bet(api, header, order)

cancel_limit_order

ManifoldAPIs.cancel_limit_order — Function

cancel_limit_order(api::AbstractAPI, header, bet_id)

Cancels a limit order based on bet ID.

https://docs.manifold.markets/api#post-v0bet

Arguments

api::AbstractAPI: API object

header: a dictionary or other compliant data structure containing authorization details

bet_id: the bet id found in the server response: response.betId

Example

The example below illustrates how to submit a limit order and cancel it based on the betId. Note that "dryRun" needs to be set to false for the example to fully work.

using ManifoldAPIs

header = Dict(
    "Authorization" => "Key authorization_key_here",
    "Content-Type" => "application/json"
)

order = Dict(
    "contractId" => "HtsxyBopv0MSywM0f0Yp",
    "amount" => 10,
    "outcome" => "YES",
    "limitProb" => .10,
    "dryRun" => true
)

api = ManifoldAPI()
response = make_bet(api, header, order)
cancel_limit_order(api, response.betId)

DataFrame Extensions

create_dataframe

ManifoldAPIs.create_dataframe — Function

create_dataframe(
    market, 
    bets;
    columns = [
        "userId",
        "outcome",
        "amount",
        "shares",
        "answerId",
        "updatedTime",
        "createdTime",
        "probAfter"
    ]
)

Converts JSON object holding market data to a DataFrame.

Arguments

market: a JSON3 object containing market information

bets: a JSON3 object containing price history

Keywords

columns: columns to include in the DataFrame. See the function signature above for default columns.

StatsPlots Extensions

plot_prices

ManifoldAPIs.plot_prices — Function

plot_prices(df::DataFrame; kwargs...)

Arguments

df::DataFrame: a DataFrame containing market information and price history

Keywords

config...: optional keyword arguments to modify the plot configuration

Example

using ManifoldAPIs
using DataFrames 
using StatsPlots 

api = ManifoldAPI()
market_slug = "supreme-court-rules-trump-tariffs-u"
market = get_market_by_slug(api, market_slug)
bets = get_bets(api; market_slug)

df = create_dataframe(market, bets)
plot_prices(df; size = (800, 400))

Utilities

buy_shares

ManifoldAPIs.buy_shares — Function

buy_shares(
    api,
    market_type; 
    target_prices,
    max_amount,
    header,
    slug,
    dry_run = false,
)

Specifiy a vector of target prices and buy shares for a total up to max_amount of Mana. Outstanding limit orders are automatically canceled.

Arguments

api::AbstractAPI: API object

market_type: the market type. Currently Single and Multiple are supported.

Keywords

target_prices: a vector of target prices for Multiple markets or a scalar for a Single market max_amount: the maximum amount of Mana to be spent

header: a dictionary or other compliant data structure containing authorization details

slug:

dry_run = false: if false, executes the order. If true, simulates the execution of an order

Example

The market used in the example below consists of 9 unlinked markets (i.e., their prices are not required to sum to 1), which can be found here: https://manifold.markets/dfish/will-the-sp-500-be-greater-than-or-cqz5Sct5Rg

The function buy_shares uses a heuristic to purchase up to max_amount worth of shares in sub-markets that have high payoffs.

Note: dry_run must be set to false in order to execute the transaction

using ManifoldAPIs

header = Dict(
    "Authorization" => "Key authorization_key_here",
    "Content-Type" => "application/json"
)

api = ManifoldAPI()
slug = "will-the-sp-500-be-greater-than-or-cqz5Sct5Rg"
max_amount = 25
target_prices = [.9, .8, .7, .6, .5, .4, .3, .2, .1]
results = buy_shares(
    api,
    Multiple; 
    target_prices,
    max_amount,
    header,
    slug,
    dry_run = true,
)

schedule

Base.schedule — Function

schedule(
    api::AbstractAPI; 
    transact,
    can_transact,
    delay = 1, 
    args_can = (),
    kwargs_can = (),
    args = (),
    kwargs = ()
)

Schedules a transaction by calling the function transact when the conditions specified in the function can_transact are satisfied.

Arguments

api = ManifoldAPI(): Manifold API object which is passed to transact and can_transact

Keywords

transact: a function that executes a transaction. Function signature: transact(api, args_can...; kwargs_can...)

can_transact: a function that specifies the conditions under which transact can be executed. Function signature: can_transact(api, args_can...; kwargs_can...)

delay = 1: delay in seconds to wait before checking can_transact

args_can = (): optional arguments for can_transact

kwargs_can = (): optional keyword arguments for can_transact

args = (): optional arguments for transact

kwargs = (): optional keyword arguments for transact

Example

using Dates 
using ManifoldAPIs

api = ManifoldAPI()

header = Dict(
    "Authorization" => "Key authorization_key_here",
    "Content-Type" => "application/json"
)

order = Dict(
    "contractId" => "HtsxyBopv0MSywM0f0Yp",
    "amount" => 10,
    "outcome" => "YES",
    "dryRun" => true
)

can_transact(api; time) = now() ≥ time
transact(api; header, order) = make_bet(api, header, order)

@async transaction = schedule(
    api; 
    transact,
    can_transact,
    kwargs = (; header, order),
    kwargs_can = (; time = now() + Second(10)),
)